fix DMC code warning C26451
See merge request Linaro/WindowsPerf/windowsperf!49

=== ChangeLog ===

* wperf-driver: fix DMC code warning C26451

There are multiple warnings in `dmc.c`:

    warning C26451: Arithmetic overflow: Using operator '+' on a 4 byte
    value and then casting the result to a 8 byte value. Cast the value
    to the wider type before calling operator '+' to avoid overflow
    (io.2).

Macro DMC_COUNTER_BASE(val) parameter is UINT8 and result of this macro
is ONLY used in UINT64 operations in DMC code.

+++ wperf-driver/dmc.h
@@ -32,7 +32,7 @@
 #include "wperf-common\macros.h"
 #include "pmu.h"
 
-#define DMC_COUNTER_BASE(val)			(0x10 + val * 0x28)
+#define DMC_COUNTER_BASE(val)			(0x10 + (UINT64)val * 0x28)
 #define DMC_COUNTER_CTL_OFFSET			(0x10)
 #define DMC_COUNTER_VAL_OFFSET			(0x20)
 #define DMC_CTL_BIT_ENABLE				(0x01 << 0)
